requirements.txt
Django==4.2.5
Pillow==9.5.0
Faker==21.0.0
django-tinymce==3.6.1
django-crispy-forms==2.1
crispy-bootstrap4
transliterate~=1.10.2
requests
markdown~=3.7

djangorestframework~=3.15.2